cloud-functions/send_email/requirements.txt
flask
google-cloud-storage
mailjet-rest